requirements.txt
streamlit>=1.21.0
pandas
pillow
pytesseract